+++ spark-bigquery-connector-common/src/main/java/com/google/cloud/spark/bigquery/SchemaConverters.java
@@ -379,18 +379,30 @@ public class SchemaConverters {
       int precision =
           Optional.ofNullable(field.getPrecision())
               .map(Long::intValue)
-              .orElse(BigQueryUtil.DEFAULT_BIG_NUMERIC_PRECISION);
+              .orElse(configuration.getBigNumericDefaultPrecision());
       if (precision > DecimalType.MAX_PRECISION()) {
         throw new IllegalArgumentException(
             String.format(
-                "BigNumeric precision is too wide (%d), Spark can only handle decimal types with max precision of %d",
+                "BigNumeric precision is too wide (%d), Spark can only handle decimal types with max precision of %d, "
+                    + "If your data is within Spark's precision, you can set it using bigNumericDefaultPrecision",
                 precision, DecimalType.MAX_PRECISION()));
+      }
+      int scale =
+          Optional.ofNullable(field.getScale())
+              .map(Long::intValue)
+              .orElse(configuration.getBigNumericDefaultScale());
+      if (scale > DecimalType.MAX_SCALE()) {
+        throw new IllegalArgumentException(
+            String.format(
+                "BigNumeric scale is too wide (%d), Spark can only handle decimal types with max scale of %d, "
+                    + "If your data is within Spark's scale, you can set it using bigNumericDefaultScale",
+                scale, DecimalType.MAX_SCALE()));
       }
       return createDecimalTypeFromNumericField(
           field,
           LegacySQLTypeName.BIGNUMERIC,
-          BigQueryUtil.DEFAULT_BIG_NUMERIC_PRECISION,
-          BigQueryUtil.DEFAULT_BIG_NUMERIC_SCALE);
+          configuration.getBigNumericDefaultPrecision(),
+          configuration.getBigNumericDefaultScale());
     } else if (LegacySQLTypeName.STRING.equals(field.getType())) {
       return DataTypes.StringType;
     } else if (LegacySQLTypeName.BOOLEAN.equals(field.getType())) {